Refactor Azure Function for unzipping large files: update README for memory constraints, enhance deployment scripts with retry logic, and improve error handling. Remove legacy code and add logging for better traceability.
