+++ src/azure-function-unzip-large-files/bicep/main.parameters.json
@@ -2,9 +2,6 @@
   "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentParameters.json#",
   "contentVersion": "1.0.0.0",
   "parameters": {
-    "storageAccountName": {
-      "value": ""
-    },
     "storageAccountSku": {
       "value": "Standard_LRS"
     },
