+++ src/azure-function-unzip-large-files/function-unzip-large-zip-files/host.json
@@ -12,5 +12,5 @@
     "id": "Microsoft.Azure.Functions.ExtensionBundle",
     "version": "[3.*, 4.0.0)"
   },
-  "functionTimeout": "00:30:00"
+  "functionTimeout": "00:10:00"
 }
