+++ src/azure-function-unzip-large-files/function-unzip-large-zip-files/requirements.txt
@@ -2,3 +2,4 @@ azure-functions
 azure-storage-blob
 pyzipper
 python-dotenv
+psutil
